+++ CHANGELOG.md
@@ -1848,3 +1848,4 @@
 - [2026-06-07 16:53] style(network): harmonize light and dark theme palette for network
 - [2026-06-07 17:18] fix(security): prevent double tap duplicate action in security
 - [2026-06-07 17:27] chore(storage): update dependencies and cleanup unused imports in storage
+- [2026-06-07 17:34] style(analytics): harmonize light and dark theme palette for analytics